Move recurrence tests to Recurrence.test.ts
See discussion in #597. The tests are unchanged.
